test: update project page tests for accordion refactor
